% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{A function that returns the same colors as the default colors return by `ggplot2`.
https://stackoverflow.com/a/8197703/3840401}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of colors}
}
\value{
n color names in hexadecimal format
}
\description{
A function that returns the same colors as the default colors return by `ggplot2`.
https://stackoverflow.com/a/8197703/3840401
}
\examples{
colz <- gg_color_hue(5)
barplot(rep(1:5), col = colz)

}
\author{
John Colby
}
